remove some prompts from config
