+++ action.yml
@@ -104,29 +104,6 @@ inputs:
       requested. Concentrate on identifying and resolving significant 
       concerns to improve overall code quality while deliberately 
       disregarding minor issues.
-  summarize_file_diff:
-    required: false
-    description: 'The prompt for summarizing a file'
-    default: |
-      GitHub pull request title: 
-      `$title` 
-
-      Description:
-      ```
-      $description
-      ```
-
-      Content of file `$filename` prior to changes:
-      ```
-      $file_content
-      ```
-
-      Diff for `$filename`:
-      ```diff
-      $file_diff
-      ```
-
-      I would like you to summarize the diff within 50 words.
   summarize:
     required: false
     description: 'The prompt for final summarization response'
@@ -162,87 +139,6 @@ inputs:
       Below the release notes, generate a short, celebratory poem about the 
       changes in this PR and add this poem as a quote (> symbol). You can 
       use emojis in the poem, where they are relevant.
-  review_file_diff:
-    required: false
-    description: 'The prompt for each file'
-    default: |
-      GitHub pull request title: 
-      `$title` 
-
-      Description:
-      ```
-      $description
-      ```
-
-      OpenAI generated notes:
-      ```
-      $release_notes
-      ```
-
-      Content of `$filename` prior to changes:
-      ```
-      $file_content
-      ```
-
-      $patches
-  comment:
-    required: false
-    description: 'Prompt for comment'
-    default: |
-      A comment was made on a GitHub pull request review for a 
-      diff hunk on file `$filename`. I would like you to follow 
-      the instructions in that comment. 
-
-      Pull request title:
-      `$title`
-
-      Description:
-      ```
-      $description
-      ```
-
-      OpenAI generated notes:
-      ```
-      $release_notes
-      ```
-
-      Content of file prior to changes:
-      ```
-      $file_content
-      ```
-
-      Entire diff:
-      ```diff
-      $file_diff
-      ```
-
-      Diff being commented on:
-      ```diff
-      $diff
-      ```
-
-      The format of a comment in the chain is:
-      `user: comment`
-
-      Comment chain (including the new comment):
-      ```
-      $comment_chain
-      ```
-
-      Please reply directly to the new comment (instead of suggesting 
-      a reply) and your reply will be posted as-is.
-
-      If the comment contains instructions/requests for you, please comply. 
-      For example, if the comment is asking you to generate documentation 
-      comments on the code, in your reply please generate the required code.
-
-      In your reply, please make sure to begin the reply by tagging the user 
-      with "@user".
-
-      The comment/request that you need to directly reply to:
-      ```
-      $comment
-      ```
 runs:
   using: 'node16'
   main: 'dist/index.js'
